backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pillow==11.0.0

